restructure test cases for simple strategy.
